No longer converting char refs, matching attributes by original values
All code related to converting character references is gone. As such,
the silly dependency on boost_thread (for the provided synchronization
mechanisms) is gone. It is not necessary to convert char refs in
selector strings because I also changed the GQAttributeSelector to do
matching against the gumbo original_name/original_value members, where
char refs and such are left intact.

Also, ran BCP against the parser header and copied the output into
third-party/include. This is so that users don't have to download and
configure a full boost release. This header only code comes from boost
release 1.59 but clients can optionally ignore this directory and
replace boost references in project files to point to their own boost
installations.