Changed GQParser::ParseIdentifier to accept escaped chars, again...
I originally made GQParser::ParseIdentifier accept escaped characters as
valid parts of an identifier. Then, I removed it. Now, I've added it
back in. :trollface: . I thought at first that this wasn't allowed, in a
selector, then I found existing selectors in the Adblock Plus Easylist
that used them. So, I've rolled back in support for it. One goal for the
tests here is to have all Easylist selectors parse with no errors.