Match system updated, a few new tests
Match system now returns a minimal structure which, if requested by the
user, can return the exact node matched by the supplied selector. This
is necessary because simply returning a boolean value was not sufficient
enough to provide accurate selector functionality.

For example, in the case of a descendant selector, a simple bool true
response would have collected and returned the ancestor to the user, not
the selected descendant. The new matching system in GQSelector and
subclasses gives accurate behavior now.

Removed the ::Find(...) methods from GQSelection. I don't believe this
mild convenience to the user is worth all the work that would need to be
done under the hood. Users can simply iterate over a match collection
and use the GQNode::Find(...) methods manually.

GQUnarySelector was named wrong this entire time. Wow I'm brutal. Anyway
I didn't do any kind of git rename, just blew it away, because I like to
blow away history like that. #yolo.