Output restructuring
Restructured the output directories for binaries and includes.

Modified Document.cpp to check to see if any children are presenting in
parsed GumboOutput and, if none are, assume invalid/bad html input and
throw.

Modified tests to safely handle streamoff returned from fstream::tellg()
to ensure its positive and safe to cast to size_t for resizing file data
containers.