Output restructuring, script updates
Output has been restructured so that includes are nested in a directory
"gq". Gumbo Parser includes are copied out also to a nested directory
"gumboparser" so includes are not top level, which collide with lots of
stuff.

Added `::Size()` method to NodeMutationCollection`.

Deleted half the scripts, consolidated into just two scripts (for
building Gumbo Parser).

Added SSE2 and AVX targets, began testing partial implementation of
matching code with these instructions, but none of this is included this
commit, because I've never worked with them and right now I'm like a
little boy with a powder keg and a box of matches.

Also, in the scripts, corrected a bug where I was linking to the static
version of the CRT. Duyee.