Staging for Gradle
Added string ref equality functionality to StrRefHash.hpp and then added
it to the unordered maps usd to within the TreeMap implementation. This
requires further testing to analyze pros/cons.

Modify gitignore to prep deps direcotry to host forked gumbo-parser as a
submodule.

Generated initial Gradle Wrapper.

Next to be done is to remove the build batch files and replace them
Gradle scripts to do the build.